requirements.txt
dict2xml==1.7.0
numpy==1.21.2
pandas==1.3.2
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
